refactor(radar): split RadarPanel/RadarFrame into mixin package
Convert widgets/frames/radar.py (509 LOC, two classes) into a package
following the match_history/timer_alert layout, with per-class mixins
so each widget composes its own handlers/properties:

- __init__.py: public API (RadarPanel, RadarFrame)
- frame.py: sys.path bootstrap, RadarPanel/RadarFrame __init__ and
  _build_ui; RadarPanel(RadarPanelHandlersMixin,
  RadarPanelPropertiesMixin, wx.Panel) and RadarFrame composed the
  same way with the Frame-side mixins
- handlers.py:
  - RadarPanelHandlersMixin -- display_radar, clear, tooltip tracking
    (_bind_tooltip_handlers, _on_list_mouse_move, _clear_tooltip),
    _populate_card_list, export/use-search click callbacks
  - RadarFrameHandlersMixin -- _load_archetypes, generate_for_archetype,
    generate/cancel click handlers, _start_generation,
    _generate_radar_worker (the threaded radar build), _export_radar
    (wx.FileDialog + atomic_write_text), _use_radar_for_search,
    _on_close (with in-progress-worker confirmation)
- properties.py:
  - RadarPanelPropertiesMixin -- _t, _format_distribution_tooltip
  - RadarFramePropertiesMixin -- _t, get_current_radar

Existing imports in widgets/handlers/app_event_handlers.py,
widgets/frames/app_frame.py, and the lazy-export map in
widgets/frames/__init__.py continue to resolve RadarFrame / RadarPanel
via the package __init__, so no call sites change.

pyproject.toml's E402 per-file-ignore now points at the new
widgets/frames/radar/frame.py where the sys.path bootstrap lives
(replacing the previous ``# ruff: noqa: E402`` file-level directive).

Refs #427 (second pass, partial).
