+++ pyproject.toml
@@ -70,6 +70,7 @@ ignore = [
 # Widget entry points insert project root into sys.path before other imports
 "widgets/frames/identify_opponent.py" = ["E402"]
 "widgets/frames/match_history/frame.py" = ["E402"]
+"widgets/frames/radar/frame.py" = ["E402"]
 "widgets/frames/timer_alert/frame.py" = ["E402"]
 "widgets/frames/top_cards/frame.py" = ["E402"]
 "widgets/frames/metagame_analysis.py" = ["E402"]
