+++ widgets/frames/radar/__init__.py
@@ -1,0 +1,7 @@
+"""Archetype radar viewer package."""
+
+from __future__ import annotations
+
+from widgets.frames.radar.frame import RadarFrame, RadarPanel
+
+__all__ = ["RadarFrame", "RadarPanel"]
